Refactor startup logic into extension methods class
Moved configuration, authentication, and service registration logic from Program.cs into new BoostrapperExtensions class. This modularizes Kestrel setup, authentication modes, MediatR/test mediator, permission services, and login/logout endpoints using extension methods. Improves code organization, readability, and maintainability by reducing Program.cs complexity and centralizing environment-specific setup.
